refactor: move operations example tests to integration
